requirements.txt
# Runtime de ANTLR necesario para CompiscriptLexer/Parser
antlr4-python3-runtime==4.13.2

streamlit==1.36.0

pytest==8.2.0

graphviz==0.20.3
pydot==1.4.2
